Check for animation metadata and hack the texture size.
Animations are handled by a filmstrip image (frames stacked beneath each other) and a `.mcmeta` file for metadata on timing etc.

So, check for this file and if it's found, force width and height to be the same throughout the texture material.

Doesn't give us animation, but it's a start (some notes on animating in comments near the code check for animation)